Réorganisation des fichiers
Quelques fichiers ont été supprimé, beaucoup ont été créé pour mieux répartir les fonctions du programme.
Les #include n'ont pas été vérifié (ça fonctionne mais certains sont peut-être plus large que nécessaire, voire inutile).
