+++ ASCIIToBinary.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void CharacterToBinary(char* bitsComposingChar, char charToConvert);
+void ConvertFileFromASCIIToBinary(char* path);
