--- StringManipulation.h
@@ -1,4 +1,0 @@
-#pragma once
-
-char* AddStringBeforeExtensionOfFileName(char* fileName, char* stringToAdd);
-void CharacterToBinary(char* bitsComposingChar, char charToConvert);
