+++ Structures/NodeHuffman.c
@@ -3,8 +3,8 @@
 #include "ListHuffman.h"
 #include "../Utilities.h"
 #include "QueueNodeHuffman.h"
+#include "../FunctionsOfStructures/ListCharAndNbOccFunctions.h"
 
-#include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 
@@ -12,8 +12,9 @@ NodeHuffman* CreateNodeHuffmanFromChar(char c)
 {
 	NodeHuffman* node = (NodeHuffman*)malloc(sizeof(NodeHuffman));
 
-	if (node != NULL)
-	{
+	if (node == NULL)
+		printf("can’t allocate memory to NodeHuffman in createNodeHuffman()\n");
+	else {
 		node->c = c;
 		node->nbOcc = 1;
 		node->left = NULL;
@@ -89,7 +90,7 @@ NodeHuffman* CreateHuffmanTree(ListCharAndNbOcc *list)
 			nextNode = (NodeHuffman*)malloc(sizeof(NodeHuffman));
 			if (nextNode != NULL)
 			{
-				nextNode->c = NULL;
+				nextNode->c = '\0';
 				nextNode->nbOcc = leftChild->nbOcc + rightChild->nbOcc;
 				nextNode->left = leftChild;
 				nextNode->right = rightChild;
@@ -123,122 +124,100 @@ NodeHuffman* NodeHuffmanFromNodeCharAndNbOcc(ListCharAndNbOcc* list)
 	return node;
 }
 
-void _WriteDictionnary(NodeHuffman* tree, const char* code, FILE *dic)
+void AddNodeHuffmanInArray(NodeHuffman** array, int arrSize, char car, int pos)
 {
-	if (tree->c != NULL)
+	if (array[pos] != NULL)
 	{
-		if (fprintf_s(dic, "%c : %s\n", tree->c, code) < 0)
-		{
-			printf("Error while writing in dico.txt\n\n");
-		}
+		for (int i = arrSize; i > pos; i--)
+			array[i] = array[i - 1];
 	}
+
+	array[pos] = CreateNodeHuffmanFromChar(car);
+}
+
+NodeHuffman** CreateArrayOfNodeHuffmanWithNbOccFromFile(FILE* f, int* sizeTab)
+{
+	NodeHuffman** tab = (NodeHuffman**)calloc(256, sizeof(NodeHuffman*));
+
+	fseek(f, 0, SEEK_SET);
+
+	if (tab == NULL)
+		printf("can’t allocate memory to array NodeHuffman in OccurOpti()\n");
 	else
 	{
-		int size = strlen(code);
-		if (size < 0)
+		char car;
+		int start, end, mid, found;
+		int tabOcc[256] = { 0 };
+		*sizeTab = 0;
+
+		while ((car = fgetc(f)) != EOF)
 		{
-			printf("error to get the size of the string \"%s\"\n\n", code);
-		}
-		else
-		{
-			size += 2;
-			char* nextCode = (char*)malloc(sizeof(char) * (size));
-			if (nextCode == NULL)
-			{
-				printf("Can't allocate memory to next code in WriteDictionnary\n\n");
-			}
-			else
-			{
-				errno_t err = strcpy_s(nextCode, size, code);
-				if (err)
-				{
-					printf("Error while copiing the old code in a new string\n\n");
+			start = 0;
+			end = *sizeTab;
+			found = 0;
+			mid = 0;
+
+			while (found == 0 && start <= end) {
+				mid = (start + end) / 2;
+
+				if (tab[mid] != NULL) {
+					if (tab[mid]->c == car)
+						found = 1;
+					else if (car > tab[mid]->c)
+						start = mid + 1;
+					else
+						end = mid - 1;
 				}
 				else
-				{
-					nextCode[size - 1] = '\0';
-					nextCode[size - 2] = '0';
-					_WriteDictionnary(tree->left, nextCode, dic);
-					nextCode[size - 2] = '1';
-					_WriteDictionnary(tree->right, nextCode, dic);
-				}
-				free(nextCode);
+					start = end + 1;
+			}
+
+			if (found != 0)
+				tab[mid]->nbOcc++;
+			else {
+				if (tab[mid] != NULL && car > tab[mid]->c)
+					mid++;
+
+				AddNodeHuffmanInArray(tab, *sizeTab, car, mid);
+				*sizeTab = *sizeTab + 1;
 			}
 		}
 	}
+	return tab;
 }
 
-void WriteDictionnary(NodeHuffman* tree)
+NodeHuffman** _CreateArrayOfNodeHuffmanWithNbOccFromFile(FILE* f, int* sizeTab)
 {
-	printf("\nWriting dictionnary...\n");
-	if (tree == NULL)
+	char car;
+	int tabOcc[256] = { 0 };
+	(*sizeTab) = 0;
+	ListCharAndNbOcc* list = NULL;
+	while ((car = fgetc(f)) != EOF)
 	{
-		printf("The dictionnary can't be write if he is empty\n");
-	}
-	else
-	{
-		FILE* dic = NULL;
-		errno_t err = fopen_s(&dic, "dico.txt", "w");
-
-		if (err || dic == NULL)
+		tabOcc[car]++;
+		if (tabOcc[car] == 1)
 		{
-			printf("Error while opening dico.txt\n");
-			PrintErrorMessageFromErrorCodeFromFile(err);
-		}
-		else
-		{
-			_WriteDictionnary(tree, "", dic);
-			printf("Dictionnary writing finished...\n");
-			fclose(dic);
+			(*sizeTab) += 1;
 		}
 	}
-}
-
-void PrintHuffmanTree(NodeHuffman* tree)
-{
-	if (tree != NULL)
+	NodeHuffman** tab = (NodeHuffman**)malloc(*sizeTab * sizeof(NodeHuffman*));
+	int i = 0;
+	car = 0;
+	while (i < *sizeTab)
 	{
-		if (tree->c == NULL)
+		while (tabOcc[car] == 0)
 		{
-			printf("/%d/ ", tree->nbOcc);
+			car++;
 		}
-		else
-		{
-			printf("%c : %d | ", tree->c, tree->nbOcc);
-		}
-
-		PrintHuffmanTree(tree->left);
-		PrintHuffmanTree(tree->right);
+		tab[i] = CreateNodeHuffmanFromChar(car);
+		tab[i]->nbOcc = tabOcc[car];
+		i++;
+		car++;
 	}
+	FreeList(list);
+	return tab;
 }
 
-void FreeHuffmanTree(NodeHuffman* tree)
-{
-	if (tree != NULL)
-	{
-		FreeHuffmanTree(tree->left);
-		FreeHuffmanTree(tree->right);
-
-		free(tree);
-	}
-}
-
-int IsSuperior(NodeHuffman** noeud1, NodeHuffman** noeud2)
-{
-	if ((*noeud1)->nbOcc < (*noeud2)->nbOcc) return -1;
-	if ((*noeud1)->nbOcc > (*noeud2)->nbOcc) return 1;
-	return 0;
-}
-
-void SortArrayByNbOcc(NodeHuffman** array, int size)
-{
-	printf("avant : %p\n", *array);
-	qsort(array, size, sizeof(NodeHuffman*), IsSuperior);
-	printf("apres : %p\n", *array);
-}
-
-
-// K
 NodeHuffman* CreateHuffmanTreeFromArray(NodeHuffman** array, int size)
 {
 	if (size == 0)
@@ -287,50 +266,6 @@ NodeHuffman* CreateHuffmanTreeFromArray(NodeHuffman** array, int size)
 			{
 				return GetDataFromQueueNodeHuffman(listNodeHuffmanTree);
 			}
-		}
-	}
-}
-
-void UpdateArrayNextHuffmanNodes(NodeHuffman** array, int size, QueueNodeHuffman* l1, QueueNodeHuffman* l2)
-{
-	for (int i = 0; i < 4; i++)
-	{
-		if (array[i] == NULL)
-		{
-			if (i < 2)
-			{
-				array[i] = GetDataFromQueueNodeHuffman(l1);
-			}
-			else
-			{
-				array[i] = GetDataFromQueueNodeHuffman(l2);
-			}
-		}
-	}
-}
-
-NodeHuffman* GetMinNodeFromArray(NodeHuffman** array, int size)
-{
-	NodeHuffman* minNode = NULL;
-
-	for (int i = 0; i < size; i++)
-	{
-		if (array[i] != NULL && (minNode == NULL || minNode->nbOcc > array[i]->nbOcc))
-		{
-			minNode = array[i];
-		}
-	}
-
-	return minNode;
-}
-
-void RemoveNodeFromArray(NodeHuffman** array, int size, NodeHuffman* nodeToRemove)
-{
-	for (int i = 0; i < size; i++)
-	{
-		if (array[i] == nodeToRemove)
-		{
-			array[i] = NULL;
 		}
 	}
 }
@@ -413,28 +348,4 @@ NodeHuffman* CreateHuffmanTreeFromDictionnaryFile(char* dicPath)
 	}
 
 	return NULL;
-}
-
-void AddNodeHuffmanInHuffmanTree(NodeHuffman** tree, char c, char* code)
-{
-	if (*tree == NULL)
-	{
-		(*tree) = CreateNodeHuffmanFromChar('\0');
-	}
-
-	if (code != NULL)
-	{
-		if (code[0] == '\0')
-		{
-			(*tree)->c = c;
-		}
-		else if (code[0] == '1')
-		{
-			AddNodeHuffmanInHuffmanTree(&((*tree)->right), c, code + 1);
-		}
-		else if (code[0] == '0')
-		{
-			AddNodeHuffmanInHuffmanTree(&((*tree)->left), c, code + 1);
-		}
-	}
 }
