+++ Utilities.h
@@ -1,5 +1,9 @@
 #pragma once
 
-#include <stdlib.h>
+#include <stdio.h>
 
 void PrintErrorMessageFromErrorCodeFromFile(errno_t err);
+
+char* AddStringBeforeExtensionOfFileName(char* fileName, char* stringToAdd);
+
+int GetNumberCharInFile(char* path);
